Replace nak_commonconf inventory variable
Replace nak_commonconf inventory variable with full inheritance
of configuration files using inherit: [filepath1, filepath2, ...]
in the YAML structure. The inheritance is recursive and the general
rule is the same as inheritance in OOP - lower levels of inheritance
have lower priority and lists are being processed from left to right
or top-down.